Fixed some bugs relating to auto-update causing crashing
Fixed some bugs relating to auto update where it would basically ping the server to check for updates, but if there is no internet the program would crash if it has auto-update turned on. Due to the built in error handling of tkinter, it would only do so if auto-update was enable and there was no internet. 

Signed-off-by: IDoUseLinux <142350850+IDoUseLinux@users.noreply.github.com>